Bunch of fixes and support for DQN

+++ apps/game1/objectLocator.cpp
@@ -11,25 +11,25 @@ using namespace std;
 
 
 void objectLocator::init() {
-    l1[SE] = 1 / sqrt(2);
-    l2[SE] = 1 / sqrt(2);
-    p1[SE] = - 1 / sqrt(2);
-    p2[SE] = 1 / sqrt(2);
+    l1[SE] = 1 / sqrtf(2);
+    l2[SE] = 1 / sqrtf(2);
+    p1[SE] = - 1 / sqrtf(2);
+    p2[SE] = 1 / sqrtf(2);
 
     l1[N] = -1;
     l2[N] = 0;
     p1[N] = 0;
     p2[N] = -1;
 
-    l1[NE] = -1 / sqrt(2);
-    l2[NE] = 1 / sqrt(2);
-    p1[NE] = -1 / sqrt(2);
-    p2[NE] = -1 / sqrt(2);
+    l1[NE] = -1 / sqrtf(2);
+    l2[NE] = 1 / sqrtf(2);
+    p1[NE] = -1 / sqrtf(2);
+    p2[NE] = -1 / sqrtf(2);
 
-    l1[NW] = -1 / sqrt(2);
-    l2[NW] = -1 / sqrt(2);
-    p1[NW] = 1 / sqrt(2);
-    p2[NW] = -1 / sqrt(2);
+    l1[NW] = -1 / sqrtf(2);
+    l2[NW] = -1 / sqrtf(2);
+    p1[NW] = 1 / sqrtf(2);
+    p2[NW] = -1 / sqrtf(2);
 
     l1[E] = 0;
     l2[E] = 1;
@@ -46,19 +46,19 @@ void objectLocator::init() {
     p1[S] = 0;
     p2[S] = 1;
 
-    l1[SW] = 1 / sqrt(2);
-    l2[SW] = -1 / sqrt(2);
-    p1[SW] = 1 / sqrt(2);
-    p2[SW] = 1 / sqrt(2);
+    l1[SW] = 1 / sqrtf(2);
+    l2[SW] = -1 / sqrtf(2);
+    p1[SW] = 1 / sqrtf(2);
+    p2[SW] = 1 / sqrtf(2);
 }
 
 void objectLocator::locateObject(int player_x, int player_y, int direction, int object_x, int object_y) {
     cout<<"Unit direction"<<direction<<endl;
-    double det = calculateDeterminant(direction);
+    float det = calculateDeterminant(direction);
     int diff1 = object_x - player_x;
     int diff2 = object_y - player_y;
-    object_l = (p2[direction] * diff1 - p1[direction] * diff2) / det;
-    object_p = (-l2[direction] * diff1 + l1[direction] * diff2) / det;
+    object_l = (p2[direction] * static_cast< float >(diff1) - p1[direction] * static_cast< float >(diff2)) / det;
+    object_p = (-l2[direction] * static_cast< float >(diff1) + l1[direction] * static_cast< float >(diff2)) / det;
     computeDistance();
     findQuadrant();
     measureUniqueAngle();
@@ -66,12 +66,12 @@ void objectLocator::locateObject(int player_x, int player_y, int direction, int 
     calculateRiskFeatures();
 }
 
-double objectLocator::calculateDeterminant(int direction) {
+float objectLocator::calculateDeterminant(int direction) {
     return l1[direction]*p2[direction] - p1[direction]*l2[direction];
 }
 
 void objectLocator::computeDistance() {
-    object_distance = sqrt(pow(object_l, 2) + pow(object_p, 2));
+    object_distance = sqrtf(powf(object_l, 2) + powf(object_p, 2));
 }
 
 void objectLocator::computeCosine() {
@@ -86,12 +86,12 @@ void objectLocator::computeSine() {
     }
 }
 
-double objectLocator::getObjectDistance() {
+float objectLocator::getObjectDistance() {
     cout<<"getObjectDistance"<<endl;
     return round_values(object_distance);
 }
 
-double objectLocator::round_values(double value) {
+float objectLocator::round_values(float value) {
     // round to 2 decimal places
     return std::round(value * 100) / 100;
 }
@@ -138,12 +138,12 @@ void objectLocator::calculateRiskFeatures() {
     cout<<"calculateRiskFeatures "<<risk_feature<<endl;
 }
 
-double objectLocator::getObjectRiskFeature() {
+float objectLocator::getObjectRiskFeature() {
     cout<<"objectLocator::getObjectRiskFeature"<<endl;
     return risk_feature;
 }
 
-double objectLocator::getObjectAngle() {
+float objectLocator::getObjectAngle() {
     cout<<"objectLocator::getObjectAngle"<<endl;
     return object_angle;
 }
